+++ apps/server/src/modules/copy-helper/copy-helper.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { CopyHelperService } from './copy-helper.service';
+
+@Module({
+	providers: [CopyHelperService],
+	exports: [CopyHelperService],
+})
+export class CopyHelperModule {}
