+++ apps/server/src/modules/files-storage-client/files-storage-client.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
-import { CopyHelperService } from '@shared/domain';
 import { LoggerModule } from '@src/core/logger';
+import { CopyHelperModule } from '@src/modules/copy-helper';
 import { CopyFilesService } from './service/copy-files.service';
 import { FilesStorageClientAdapterService } from './service/files-storage-client.service';
 import { FilesStorageProducer } from './service/files-storage.producer';
 
 @Module({
-	imports: [LoggerModule],
+	imports: [LoggerModule, CopyHelperModule],
 	controllers: [],
-	providers: [FilesStorageClientAdapterService, CopyFilesService, FilesStorageProducer, CopyHelperService],
+	providers: [FilesStorageClientAdapterService, CopyFilesService, FilesStorageProducer],
 	exports: [FilesStorageClientAdapterService, CopyFilesService],
 })
 export class FilesStorageClientModule {}
