+++ apps/server/src/modules/files-storage-client/service/copy-files.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@nestjs/common';
-import { CopyElementType, CopyHelperService, CopyStatus, CopyStatusEnum, EntityId } from '@shared/domain';
+import { EntityId } from '@shared/domain';
+import { CopyElementType, CopyHelperService, CopyStatus, CopyStatusEnum } from '@src/modules/copy-helper';
 import { CopyFileDto } from '../dto';
 import { EntityWithEmbeddedFiles } from '../interfaces';
-import { FileParamBuilder, CopyFilesOfParentParamBuilder } from '../mapper';
+import { CopyFilesOfParentParamBuilder, FileParamBuilder } from '../mapper';
 import { FilesStorageClientAdapterService } from './files-storage-client.service';
 
 // TODO  missing FileCopyParams  ...passing user instead of userId
@@ -50,11 +51,12 @@ export class CopyFilesService {
 
 	private deriveCopyStatus(fileDtos: CopyFileDto[]): CopyStatus {
 		const fileStatuses: CopyStatus[] = fileDtos.map(({ sourceId, id, name }) => {
-			return {
+			const result = {
 				type: CopyElementType.FILE,
 				status: id ? CopyStatusEnum.SUCCESS : CopyStatusEnum.FAIL,
 				title: name ?? `(old fileid: ${sourceId})`,
 			};
+			return result;
 		});
 
 		const fileGroupStatus = {
