+++ apps/server/src/modules/learnroom/controller/dto/index.ts
@@ -1,5 +1,4 @@
 export * from './board';
-export * from './copy.response';
 export * from './course-metadata.response';
 export * from './dashboard.response';
 export * from './dashboard.url.params';
