+++ apps/server/src/modules/learnroom/learnroom.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { CopyHelperService, EtherpadService, NexboardService } from '@shared/domain';
+import { EtherpadService, NexboardService } from '@shared/domain';
 import { FileLegacyService } from '@shared/domain/service/file-legacy.service';
 import { FeathersServiceProvider } from '@shared/infra/feathers';
 import {
@@ -13,6 +13,7 @@ import {
 } from '@shared/repo';
 import { Logger } from '@src/core/logger';
 import { AuthorizationModule } from '../authorization';
+import { CopyHelperModule } from '../copy-helper/copy-helper.module';
 import { FilesStorageClientModule } from '../files-storage-client';
 import { LessonModule } from '../lesson';
 import { TaskModule } from '../task';
@@ -21,12 +22,12 @@ import { DashboardController } from './controller/dashboard.controller';
 import { RoomsController } from './controller/rooms.controller';
 import { RoomBoardResponseMapper } from './mapper/room-board-response.mapper';
 import { BoardCopyService } from './service/board-copy.service';
+import { CommonCartridgeExportService } from './service/common-cartridge-export.service';
 import { CourseCopyService } from './service/course-copy.service';
+import { CourseService } from './service/course.service';
 import { LessonCopyService } from './service/lesson-copy.service';
 import { MetadataLoader } from './service/metadata-loader.service';
 import { RoomsService } from './service/rooms.service';
-import { CommonCartridgeExportService } from './service/common-cartridge-export.service';
-import { CourseService } from './service/course.service';
 import { CourseCopyUC } from './uc/course-copy.uc';
 import { CourseExportUc } from './uc/course-export.uc';
 import { CourseUc } from './uc/course.uc';
@@ -35,10 +36,9 @@ import { LessonCopyUC } from './uc/lesson-copy.uc';
 import { RoomBoardDTOFactory } from './uc/room-board-dto.factory';
 import { RoomsAuthorisationService } from './uc/rooms.authorisation.service';
 import { RoomsUc } from './uc/rooms.uc';
-import { TaskCopyService } from './service';
 
 @Module({
-	imports: [AuthorizationModule, FilesStorageClientModule, LessonModule, TaskModule],
+	imports: [AuthorizationModule, FilesStorageClientModule, LessonModule, TaskModule, TaskModule, CopyHelperModule],
 	controllers: [DashboardController, CourseController, RoomsController],
 	providers: [
 		DashboardUc,
@@ -60,8 +60,6 @@ import { TaskCopyService } from './service';
 		BoardCopyService,
 		LessonCopyService,
 		LessonCopyUC,
-		TaskCopyService,
-		CopyHelperService,
 		CourseCopyService,
 		CourseCopyUC,
 		RoomsService,
