+++ apps/server/src/modules/learnroom/service/index.ts
@@ -1,4 +1,3 @@
 export * from './board-copy.service';
 export * from './course-copy.service';
 export * from './lesson-copy.service';
-export * from './task-copy.service';
