+++ apps/server/src/modules/learnroom/types/lesson-copy-parent.params.ts
@@ -1,0 +1,6 @@
+import { EntityId } from '@shared/domain';
+
+export type LessonCopyParentParams = {
+	courseId?: EntityId;
+	userId: string;
+};
