+++ apps/server/src/modules/learnroom/types/lesson-copy.params.ts
@@ -1,0 +1,8 @@
+import { Course, Lesson, User } from '@shared/domain';
+
+export type LessonCopyParams = {
+	originalLesson: Lesson;
+	destinationCourse: Course;
+	user: User;
+	copyName?: string;
+};
