+++ apps/server/src/modules/learnroom/uc/course-copy.uc.ts
@@ -1,8 +1,9 @@
 import { Configuration } from '@hpi-schul-cloud/commons';
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
-import { Actions, CopyStatus, EntityId, Permission } from '@shared/domain';
+import { Actions, EntityId, Permission } from '@shared/domain';
 import { AuthorizationService } from '@src/modules/authorization';
 import { AllowedAuthorizationEntityType } from '@src/modules/authorization/interfaces';
+import { CopyStatus } from '@src/modules/copy-helper';
 import { CourseCopyService } from '../service';
 
 @Injectable()
