+++ apps/server/src/modules/task/index.ts
@@ -1,1 +1,3 @@
+export * from './service/task-copy.service';
+export * from './service/task.service';
 export * from './task.module';
