+++ apps/server/src/modules/task/service/index.ts
@@ -1,2 +1,3 @@
 export * from './submission.service';
+export * from './task-copy.service';
 export * from './task.service';
