+++ apps/server/src/modules/task/service/task-copy.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { CopyHelperService } from '@shared/domain';
 import { Course, Lesson, Task, User } from '@shared/domain/entity';
-import { CopyElementType, CopyStatus, CopyStatusEnum } from '@shared/domain/types';
 import { TaskRepo } from '@shared/repo';
+import { CopyElementType, CopyHelperService, CopyStatus, CopyStatusEnum } from '@src/modules/copy-helper';
 import { CopyFilesService } from '@src/modules/files-storage-client';
 import { FileUrlReplacement } from '@src/modules/files-storage-client/service/copy-files.service';
