+++ apps/server/src/modules/task/uc/index.ts
@@ -1,2 +1,3 @@
-export * from './task.uc';
 export * from './submission.uc';
+export * from './task-copy.uc';
+export * from './task.uc';
