+++ apps/server/src/shared/domain/service/index.ts
@@ -1,4 +1,3 @@
-export * from './copy-helper.service';
 export * from './etherpad.service';
 export * from './nexboard.service';
 export * from './permission.service';
