+++ apps/server/src/shared/domain/types/index.ts
@@ -1,11 +1,10 @@
-export * from './copy.types';
 export * from './counted';
 export * from './entity-id';
-export * from './learnroom.types';
-export * from './input-format.types';
 export * from './importuser.types';
+export * from './input-format.types';
+export * from './learnroom.types';
 export * from './news.types';
 export * from './richtext.types';
-export * from './task.types';
 export * from './system.type';
+export * from './task.types';
 export * from './value-of';
